requirements.txt
torch>=2.1.0
numpy>=1.23,<2.0
scipy>=1.10
scikit-learn>=1.3,<1.6
PyYAML>=6.0
